+++ packages/react-magma-docs/gatsby-config.js
@@ -56,6 +56,7 @@ module.exports = {
             resolve: 'gatsby-remark-images',
             options: {
               maxWidth: 1035,
+              linkImagesToOriginal: false,
             },
           },
         ],