+++ packages/react-magma-dom/src/components/IndeterminateCheckbox/index.tsx
@@ -11,6 +11,9 @@ import { StyledLabel } from '../SelectionControls/StyledLabel';
 import { StyledContainer } from '../SelectionControls/StyledContainer';
 import styled from '@emotion/styled';
 import { useGenerateId, Omit } from '../../utils';
+import { VisuallyHidden } from '../VisuallyHidden';
+import { Announce } from '../Announce';
+import { I18nContext } from '../../i18n';
 
 export interface IndeterminateCheckboxProps
   extends Omit<CheckboxProps, 'checked'> {
@@ -70,6 +73,7 @@ export const IndeterminateCheckbox: React.FunctionComponent<IndeterminateCheckbo
     }
 
     const theme = React.useContext(ThemeContext);
+    const i18n = React.useContext(I18nContext);
 
     const {
       color,
@@ -84,6 +88,39 @@ export const IndeterminateCheckbox: React.FunctionComponent<IndeterminateCheckbo
       testId,
       ...other
     } = props;
+
+    const isIndeterminate = status === 'indeterminate';
+    const isUnchecked = status === 'unchecked';
+
+    function replaceLabelTextForAnnounceText(baseAnnounceText) {
+      return baseAnnounceText.replace(
+        /\{labelText\}/g,
+        getStringifiedLabelText(labelText)
+      );
+    }
+
+    function getStringifiedLabelText(node) {
+      if (['string', 'number'].includes(typeof node)) return node;
+      if (node instanceof Array)
+        return node.map(getStringifiedLabelText).join('');
+      if (typeof node === 'object' && node)
+        return getStringifiedLabelText(node.props.children);
+    }
+
+    const showAnnounce = isChecked || isIndeterminate || isUnchecked;
+    const announceText = isChecked
+      ? replaceLabelTextForAnnounceText(
+          i18n.indeterminateCheckbox.isCheckedAnnounce
+        )
+      : isIndeterminate
+      ? replaceLabelTextForAnnounceText(
+          i18n.indeterminateCheckbox.isIndeterminateAnnounce
+        )
+      : isUnchecked
+      ? replaceLabelTextForAnnounceText(
+          i18n.indeterminateCheckbox.isUncheckedAnnounce
+        )
+      : '';
 
     return (
       <StyledContainer style={containerStyle}>
@@ -102,12 +139,12 @@ export const IndeterminateCheckbox: React.FunctionComponent<IndeterminateCheckbo
             checked={isChecked}
             color={color ? color : ''}
             disabled={disabled}
-            isIndeterminate={status === 'indeterminate'}
+            isIndeterminate={isIndeterminate}
             isInverse={isInverse}
             style={inputStyle}
             theme={theme}
           >
-            {status === 'indeterminate' && (
+            {isIndeterminate && (
               <IndeterminateIcon
                 data-testid="indeterminateIcon"
                 color={color ? color : ''}
@@ -123,6 +160,9 @@ export const IndeterminateCheckbox: React.FunctionComponent<IndeterminateCheckbo
             labelText
           )}
         </StyledLabel>
+        <Announce>
+          {showAnnounce && <VisuallyHidden>{announceText}</VisuallyHidden>}
+        </Announce>
       </StyledContainer>
     );
   }