+++ packages/react-magma-legacy-selects/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.81](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.80...v2.0.0-beta.81) (2020-09-23)
+
+**Note:** Version bump only for package react-magma-legacy-selects
+
+
+
+
+
+# [2.0.0-beta.80](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.79...v2.0.0-beta.80) (2020-09-23)
+
+**Note:** Version bump only for package react-magma-legacy-selects
+
+
+
+
+
+# [2.0.0-beta.79](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.78...v2.0.0-beta.79) (2020-09-23)
+
+
+### Bug Fixes
+
+* **docs:** fix icon dependency ([387e639](http://stash.cengage.com:7999/FRONT/react-magma/commits/387e6399783f4f5f0d6dcdd0179c4cbf06945d71))
+
+
+
+
+
 # [2.0.0-beta.78](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.77...v2.0.0-beta.78) (2020-09-18)
 
 **Note:** Version bump only for package react-magma-legacy-selects